fksn
find the k-th smallest number in two sorted array, the type of elements
are unsigned integers. There are two implementations in the source file,
one of them has the time complexity of O(k) and the another has O(lgm +
lgn), where m and n is the length of two arrays.